Reimplement Sharp Hook
